requirements.txt
absl-py==2.2.1
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
blessed==1.20.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comm==0.2.2
# Editable install with no version control (conch==0.1.0)
-e /data/zrf/VLSA-main/other_MIL_code/ViLa-MIL-main/models/CONCH-main
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
diffusers==0.33.0
docker-pycreds==0.4.0
ecos==2.0.14
einops==0.8.1
einops-exts==0.0.4
et_xmlfile==2.0.0
executing==2.2.0
faiss==1.11.0
filelock==3.18.0
fonttools==4.56.0
formulaic==1.1.1
frozenlist==1.6.0
fsspec==2025.3.2
ftfy==6.3.1
future==1.0.0
geopandas==1.1.0
gitdb==4.0.12
GitPython==3.1.44
gpustat==1.1.1
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
interface-meta==1.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lifelines==0.30.0
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
ml_collections==1.1.0
mpmath==1.3.0
multidict==6.4.3
nest-asyncio==1.6.0
networkx==3.4.2
nmslib @ file:///home/conda/feedstock_root/build_artifacts/nmslib_1725861669981/work/python_bindings
numba==0.61.0
numexpr==2.10.2
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=b8e59fcb10ab071cbeded47aed247ddfa1c902d595848c27f82021340e226134
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
nystrom-attention==0.0.14
openai-clip==1.0.1
opencv-python==4.11.0.86
openpyxl==3.1.5
openslide-bin==4.0.0.6
openslide-python==1.4.1
osqp==0.6.7.post3
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
POT==0.9.5
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663123172/work
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11 @ file:///D:/bld/pybind11-split_1730237329882/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1730237328107/work
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyMuPDF==1.26.4
pynndescent==0.5.13
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
qdldl==0.1.7.post5
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.15.2
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==75.8.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torch-geometric==2.6.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
triton==3.2.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
wandb==0.19.9
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xformers==0.0.29.post3
yarl==1.20.0
zipp==3.21.0
